// backend/src/types/enums.ts
export enum Role {
  SUPER_ADMIN = "SUPER_ADMIN",
  ADMIN = "ADMIN",
  GURU = "GURU",
  STAFF = "STAFF"
}

export enum Gender {
  L = "L",
  P = "P"
}

export enum SubjectGroup {
  KELOMPOK_A = "KELOMPOK_A", // Nasional
  KELOMPOK_B = "KELOMPOK_B", // Muatan Lokal
  KELOMPOK_C = "KELOMPOK_C"  // Ciri Khas Madrasah
}

export enum SemesterType {
  ODD = "ODD",
  EVEN = "EVEN"
}